move ChecklistLayout to model
